DEAR-90 refactor happiness radiobuttons
